+++ simpleble/src/frontends/base/Advanced.cpp
@@ -1,0 +1,26 @@
+#include "simpleble/Advanced.h"
+
+#if defined(_WIN32)
+    // Windows code
+
+#endif
+
+#if TARGET_OS_OSX
+    // macOS code
+
+#endif
+
+#if TARGET_OS_IOS
+    // iOS code
+
+#endif
+
+#if defined(__ANDROID__)
+    // Android code
+
+#endif
+
+#if defined(__linux__) && !defined(__ANDROID__)
+    // Linux code
+
+#endif